[refac] move dtos to another package
